Dockerize project based on Alpine Linux
We use a multi-stage dockerfile that builds the application in
the first stage and copies over the binary to the second one
where only the necessary dependencies are installed.

Alpine has all the dependencies in it's main repositories, only
log4cpp had to be installed from the testing one.

The resulting docker image is quite small at 116MB,
a single-binary image would be even smaller but some of the
libraries are not easily linked statically.